financeGuard/requirements.txt
flask[async]
flask-cors
flask-migrate
flask-mail
flask-socketio
python-socketio
python-engineio
python-dotenv
pytest
pytest-cov
pyjwt
flask-sqlalchemy
sqlalchemy>=2.0.0
aiomysql
pymysql
pandas
numpy
scikit-learn
openpyxl
python-dateutil
schedule
# Database dependencies
mysqlclient>=2.0.0
# Real-time & async
APScheduler
python-dateutil
